requirements/test.txt
pytest
pytest-django
pytest-pythonpath
pytest-cov
mock
